remove deformable_detr
