--- src/optimum/rbln/transformers/models/deformable_detr/modeling_deformable_detr.py
@@ -1,36 +1,0 @@
-# Copyright 2025 Rebellions Inc. All rights reserved.
-
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at:
-
-#     http://www.apache.org/licenses/LICENSE-2.0
-
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from transformers.models.deformable_detr.modeling_deformable_detr import DeformableDetrObjectDetectionOutput
-
-from ...modeling_generic import RBLNModelForObjectDetection
-from ....utils.logging import get_logger
-
-logger = get_logger(__name__)
-
-class RBLNDeformableDetrForObjectDetection(RBLNModelForObjectDetection):
-    def forward(self, *args, **kwargs):
-        outputs = super().forward(*args, **kwargs)
-
-        logits = outputs[0]
-        pred_boxes = outputs[1]
-        last_hidden_state = outputs[2]
-        encoder_last_hidden_state = outputs[3]
-
-        return DeformableDetrObjectDetectionOutput(
-            logits=logits,
-            pred_boxes=pred_boxes,
-            last_hidden_state=last_hidden_state,
-            encoder_last_hidden_state=encoder_last_hidden_state,
-        )
